build: fix mingw format checking via the gnu archetype, not the macros
mingw's R_ATTR_PRINTF/SCANF resolved to the __printf__/__scanf__
archetypes, which check against the MS C runtime and reject the C99
hh/ll length modifiers our format macros emit (MSVC's own compiler does
no such checking, so this only surfaced under mingw). The fix belongs at
the archetype, not the macros -- the macros are correct and dual-use
(hh is the only scanf-correct modifier for an 8-bit target):

 - On mingw, use the gnu_printf/gnu_scanf archetypes (C99-aware) and
   define __USE_MINGW_ANSI_STDIO so the linked runtime matches the
   checker.
 - The gnu checker rejected RINTMAX/RINTPTR's MS-style "I64", which is a
   legacy spelling for an 8-byte long long that the size_t branch already
   spells "ll". Make the 8-byte ptr/intmax branch consistent ("ll"); the
   UCRT (VS2015+, already required for %hh) and mingw both accept it.
 - rtest's SEH filter logged the DWORD ExceptionCode with %u; it is
   unsigned long, so use %lu.

With the formats checked correctly, drop -Dwerror=false from the
windows-mingw CI job so it runs at warning_level=2 + werror like the
MSVC jobs.

+++ rlib/rtest.c
@@ -846,7 +846,7 @@ _r_test_win32_exception_filter (PEXCEPTION_POINTERS ep)
 
   if (ep->ExceptionRecord->ExceptionCode != EXCEPTION_STACK_OVERFLOW) {
     r_log (R_LOG_CAT_DEFAULT, R_LOG_LEVEL_ERROR, errfile, errline, errfunc,
-        "exception code: %u", ep->ExceptionRecord->ExceptionCode);
+        "exception code: %lu", (unsigned long) ep->ExceptionRecord->ExceptionCode);
   }
 
   _r_test_win32_err_handler (SIGSEGV);
